+++ lib/src/stream/enum_auth_status.dart
@@ -1,0 +1,6 @@
+enum EnumAuthStatus {
+  signedIn,
+  signedOut,
+  // If the sign-in check has not yet been performed.
+  unknown,
+}